# Improve the performance of RollingWindow.GetEnumerator