@starteleport 2024-12-05
Thanks @jaredbroad! I wonder whether it's worth adding a performance benchmark for this case? There are not so many benchmarks now in the code base, hence the question.