@Martin-Molinero 2024-12-05
Nice work @starteleport! Could you please add some unit tests at `RollingWindowTests` for the enumeration case, there aren't many really, I see only 1 `EnumeratesAsExpected`, think would be nice to add cases where we add beyond the rolling size and assert it works as expected. Your indexing logic for list is different than the one used in `[]` => would be nice to create maybe some private method to get the index for `i` and have it in a single location, avoid the duplication in get/set/enumerate