@starteleport 2024-12-05
> would be nice to create maybe some private method to get the index for i and have it in a single location, avoid the duplication in get/set/enumerate & potential bugs

@Martin-Molinero, I would be happy to do so but sadly this method wouldn't get inlined even if I use `[MethodImpl(MethodImplOptions.AggressiveInlining)]`. That's why I left it as-is. I will try one more time though, I have an idea.