@starteleport 2024-12-05
@Martin-Molinero done, I reverted custom indexing logic and extended `EnumeratesAsExpected` test case. Think it's enough. WDYT?